+++ frontend/src/pages/MainPage/mainPage.scss
@@ -2,7 +2,7 @@
 @import '../../styles/utils';
 
 .main__banner {
-  background: url('../../assets/images/avatar.png');
+  background: url('../../assets/images/background.svg');
   background-size: 400px;
   background-repeat: no-repeat;
   background-position: 90% 150%;
